Atari++ 1.84
This version is fixing mostly issues for building Atari++ with latest g++ compilers, and running the emulator on the Raspi.
The alsa sound interface of the Raspi seems to be somewhat incomplete, thus some system calls fail on it, but with some minor tweaks and workarounds, you can get it working.