Atari++ 1.85
This release fixes a couple of configuration and error handling issues where, upon a configuration error, you could not reach the menu anymore. It also includes an improved configuration script and makefile that includes additional tests for compiler switches. This release of Atari++ should be again fine to compile with the latest g++ and clang compiler.